requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.12.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
av==16.1.0
babel==2.18.0
bangla==0.0.5
beautifulsoup4==4.14.3
blinker==1.9.0
blis==1.3.3
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2022.12.7
cffi==2.0.0
charset-normalizer==2.1.1
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cryptography==46.0.5
ctranslate2==4.7.1
cycler==0.12.1
cymem==2.0.13
Cython==3.2.4
dateparser==1.1.8
decorator==5.2.1
deepface==0.0.99
distro==1.9.0
docopt==0.6.2
einops==0.8.2
elevenlabs==2.37.0
encodec==0.1.1
exceptiongroup==1.3.1
facenet-pytorch==2.5.3
faster-whisper==1.2.1
ffmpeg-python==0.2.0
filelock==3.20.0
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
g2pkk==0.1.2
gast==0.4.0
gdown==5.2.1
google-auth==2.48.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
groq==1.0.0
grpcio==1.78.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gunicorn==25.1.0
h11==0.16.0
h5py==3.15.1
hangul-romanize==0.1.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.4
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.3
jsonlines==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.0
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
llvmlite==0.46.0
lz4==4.4.5
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.2
mtcnn==1.0.0
multidict==6.7.1
murmurhash==1.0.15
namex==0.1.0
networkx==3.4.2
nltk==3.9.3
num2words==0.5.14
numba==0.64.0
numpy==1.22.0
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.19.0.56
oauthlib==3.3.1
onnxruntime==1.23.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==1.5.3
pillow==12.0.0
platformdirs==4.9.2
playsound==1.2.2
pooch==1.9.0
preshed==3.0.12
proglog==0.1.12
propcache==0.4.1
protobuf==3.19.6
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynndescent==0.6.0
pyparsing==3.3.2
pypinyin==0.55.0
pyreadline3==3.5.4
pysbd==0.3.4
PySocks==1.7.1
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.28.1
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.11.4
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.8.3
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
SudachiDict-core==20260116
SudachiPy==0.6.10
sympy==1.13.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
tf_keras==2.20.1
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.15.2
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.3
trainer==0.0.36
transformers==4.36.2
TTS==0.22.0
typeguard==4.5.1
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==1.26.13
wasabi==1.1.3
weasel==0.4.3
webrtcvad-wheels==2.0.14
websockets==16.0
Werkzeug==3.1.6
wrapt==2.1.1
yarl==1.23.0
